vectordb/requirements.txt
fastapi
uvicorn[standard]
pillow
torch
sentence-transformers
chromadb
python-multipart

